+++ CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Methods `scan` and `scanInternet`.
 
+### 🎉 Added
+
+- Method `scanResult`.
+
 ## [3.1.5] - 2019-10-25
 
 ### ✏️ Changed